Inverted Index Construction
In this assignment the goal was to build an efficient boolean retrieval system for English corpora. <br />
The documents in the corpora are in the XML fragment form. The document collection is inverted and an on-disk inverted index built, consisting of a dictionary file and a single file of all postings lists.<br />
Then to make the storage efficient 4 different postings list compression models are implemented and corresponding to that the decompression model of each of them.<br />
The system then responds to both the Single keyword retrieval and Multi-keyword retrieval efficiently.<br />