Prior Ranking of Documents
The goal of this assignment is to simply compute prior similarity-based authority of documents and qualitatively analyse the results. 
We use of the famous "20news-groups" dataset in this. 
First the similarity graph between each pair of documents in the collection computes using two similarity algorithms which are Cosine Similarity and Jaccard Similarity.
Then on this similarity graph, the PageRank scores of all documents is computed using the Scikit-Network module.